This advancement gets the project ready to start integrating the library to show characters on the screen. The project is ready now
