+++ NodeGraphics/ePaperDriver/Debug/Driver/subdir_vars.mk
@@ -4,21 +4,33 @@
 
 # Add inputs and outputs from these tool invocations to the build variables 
 C_SRCS += \
-../Driver/SPIinit.c 
+../Driver/GPIOInit.c \
+../Driver/SPIinit.c \
+../Driver/ePaper.c 
 
 C_DEPS += \
-./Driver/SPIinit.d 
+./Driver/GPIOInit.d \
+./Driver/SPIinit.d \
+./Driver/ePaper.d 
 
 OBJS += \
-./Driver/SPIinit.obj 
+./Driver/GPIOInit.obj \
+./Driver/SPIinit.obj \
+./Driver/ePaper.obj 
 
 OBJS__QUOTED += \
-"Driver/SPIinit.obj" 
+"Driver/GPIOInit.obj" \
+"Driver/SPIinit.obj" \
+"Driver/ePaper.obj" 
 
 C_DEPS__QUOTED += \
-"Driver/SPIinit.d" 
+"Driver/GPIOInit.d" \
+"Driver/SPIinit.d" \
+"Driver/ePaper.d" 
 
 C_SRCS__QUOTED += \
-"../Driver/SPIinit.c" 
+"../Driver/GPIOInit.c" \
+"../Driver/SPIinit.c" \
+"../Driver/ePaper.c"
