+++ NodeGraphics/ePaperDriver/Debug/ccsObjs.opt
@@ -1,1 +1,1 @@
-"./main.obj" "./startup_msp432p401r_ccs.obj" "./system_msp432p401r.obj" "./Driver/SPIinit.obj" "../msp432p401r.cmd" -l"ti/display/lib/display.aem4f" -l"ti/grlib/lib/ccs/m4f/grlib.a" -l"third_party/spiffs/lib/ccs/m4f/spiffs.a" -l"ti/drivers/lib/drivers_msp432p401x.aem4f" -l"third_party/fatfs/lib/ccs/m4f/fatfs.a" -l"ti/devices/msp432p4xx/driverlib/ccs/msp432p4xx_driverlib.lib" -llibc.a 
+"./main.obj" "./startup_msp432p401r_ccs.obj" "./system_msp432p401r.obj" "./Driver/GPIOInit.obj" "./Driver/SPIinit.obj" "./Driver/ePaper.obj" "../msp432p401r.cmd" -l"ti/display/lib/display.aem4f" -l"ti/grlib/lib/ccs/m4f/grlib.a" -l"third_party/spiffs/lib/ccs/m4f/spiffs.a" -l"ti/drivers/lib/drivers_msp432p401x.aem4f" -l"third_party/fatfs/lib/ccs/m4f/fatfs.a" -l"ti/devices/msp432p4xx/driverlib/ccs/msp432p4xx_driverlib.lib" -llibc.a
