+++ NodeGraphics/ePaperDriver/Debug/makefile
@@ -11,7 +11,9 @@ ORDERED_OBJS += \
 "./main.obj" \
 "./startup_msp432p401r_ccs.obj" \
 "./system_msp432p401r.obj" \
+"./Driver/GPIOInit.obj" \
 "./Driver/SPIinit.obj" \
+"./Driver/ePaper.obj" \
 "../msp432p401r.cmd" \
 $(GEN_CMDS__FLAG) \
 -l"ti/display/lib/display.aem4f" \
@@ -163,8 +165,8 @@ ePaperDriver.hex: $(EXE_OUTPUTS)
 # Other Targets
 clean:
 	-$(RM) $(BIN_OUTPUTS__QUOTED)$(EXE_OUTPUTS__QUOTED)
-	-$(RM) "main.obj" "startup_msp432p401r_ccs.obj" "system_msp432p401r.obj" "Driver/SPIinit.obj" 
-	-$(RM) "main.d" "startup_msp432p401r_ccs.d" "system_msp432p401r.d" "Driver/SPIinit.d" 
+	-$(RM) "main.obj" "startup_msp432p401r_ccs.obj" "system_msp432p401r.obj" "Driver/GPIOInit.obj" "Driver/SPIinit.obj" "Driver/ePaper.obj" 
+	-$(RM) "main.d" "startup_msp432p401r_ccs.d" "system_msp432p401r.d" "Driver/GPIOInit.d" "Driver/SPIinit.d" "Driver/ePaper.d" 
 	-@echo 'Finished clean'
 	-@echo ' '
