+++ NodeGraphics/ePaperDriver/Driver/GPIOInit.h
@@ -1,0 +1,15 @@
+
+#include "pinDefinintions.h"
+
+
+
+void GPIO_Init(void);
+
+
+void digitalLow(uint8_t pin);
+
+
+void digitalHigh(uint8_t pin);
+
+
+uint8_t digitalRead(uint8_t pin);
