+++ NodeGraphics/ePaperDriver/Driver/SPIinit.h
@@ -9,3 +9,12 @@
 
 
 void SPI_Init(void);
+
+
+void writeCMD(uint8_t command);
+
+
+void writeData(uint8_t data);
+
+
+void SPI_transmit(uint8_t data);
