+++ NodeGraphics/ePaperDriver/Driver/ePaper.h
@@ -1,0 +1,18 @@
+
+#include "GPIOInit.h"
+#include "SPIinit.h"
+
+#define HRES 128
+#define VRES 296
+
+void powerON(void);
+
+void powerOff(void);
+
+void initEPD(void);
+
+void turnWhite(void);
+
+void turnBlack(void);
+
+void turnYellow(void);
